Update README.md
When the user first enters the website, they are prompted to choose whether they want to borrow a book, return a book, or view the available books. The user selects the option they want. If they choose to borrow or return, they will need to enter the ISBN number to perform the action on the desired book, with a maximum of two borrowed books allowed per user.